% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_data.R
\name{data_read_csv}
\alias{data_read_csv}
\title{Read a csv file}
\usage{
data_read_csv(filename)
}
\arguments{
\item{filename}{character string}
}
\value{
dataframe of whatever CSV file it read.
}
\description{
Read a csv file
}
