% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_data.R
\name{file_selection}
\alias{file_selection}
\title{Make a File Selection Object}
\usage{
file_selection(item_id, label = "Which File?", ...)
}
\arguments{
\item{item_id}{character string NS id for UI}

\item{label}{character string name on label}

\item{...}{additional arguments if you want them}
}
\value{
file selection object
}
\description{
Make a File Selection Object
}
